% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Veritabanı bağlantısını kurar ve bağlantı nesnesini döndürür}
\usage{
connect_to_database(db_path)
}
\arguments{
\item{db_path}{Veritabanı dosyasının yolunu belirtir.}
}
\value{
Veritabanı bağlantısı nesnesi.
}
\description{
Veritabanı bağlantısını kurar ve bağlantı nesnesini döndürür
}
