% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{fetch_and_fill_na}
\alias{fetch_and_fill_na}
\title{Veritabanından veri çeken ve eksik değerleri bir önceki ve bir sonraki satırın ortalaması ile dolduran fonksiyon}
\usage{
fetch_and_fill_na(db_path, table_name)

fetch_and_fill_na(db_path, table_name)
}
\arguments{
\item{db_path}{SQLite veritabanı dosyasının yolu.}

\item{table_name}{Çekilecek veritabanı tablosunun adı.}
}
\value{
Düzeltilmiş veri çerçevesi.

Düzeltilmiş veri çerçevesi.
}
\description{
Veritabanından veri çeken ve eksik değerleri bir önceki ve bir sonraki satırın ortalaması ile dolduran fonksiyon

Veritabanından veri çeken ve eksik değerleri bir önceki ve bir sonraki satırın ortalaması ile dolduran fonksiyon
}
