% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{import_csv_to_sqlite}
\alias{import_csv_to_sqlite}
\title{CSV dosyasını SQLite veritabanına aktarır}
\usage{
import_csv_to_sqlite(db_path, csv_file, table_name)
}
\arguments{
\item{db_path}{Veritabanı dosyasının yolunu belirtir.}

\item{csv_file}{CSV dosyasının yolu.}

\item{table_name}{SQLite'daki tablo ismi.}
}
\description{
CSV dosyasını SQLite veritabanına aktarır
}
